requirements.txt
Django==1.11.7
Pillow==4.3.0

